% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLocEvent.R
\name{joinLocEvent}
\alias{joinLocEvent}
\title{joinLocEvent: merges Location and Event level data with options for filtering.}
\usage{
joinLocEvent(
  park = "all",
  from = 2007,
  to = 2019,
  QAQC = FALSE,
  rejected = FALSE,
  panels = 1:4,
  locType = "VS",
  output = "short",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one park at a time. Acceptable options are:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"APCO"}{Appomattox Court House NHP only}
\item{"BOWA"}{Booker T. Washington NM only}
\item{"COLO"}{Colonial NHP only}
\item{"FRSP"}{Fredericksburg & Spotsylvania NMP only}
\item{"GETT"}{Gettysburg NMP only}
\item{"GEWA"}{George Washington Birthplace NM only}
\item{"HOFU"}{Hopewell Furnace NHS only}
\item{"PETE"}{Petersburg NBP only}
\item{"RICH"}{Richmond NB only}
\item{"SAHI"}{Sagamore Hill NHS only}
\item{"THST"}{Thomas Stone NHS only}
\item{"VAFO"}{Valley Forge NHP only}}}

\item{from}{Year to start analysis, ranging from 2007-2019}

\item{to}{Year to stop analysis, ranging from 2007-2019}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{rejected}{Allows you to remove (FALSE) or include (TRUE) rejected plots.
\describe{
\item{FALSE}{Default. Only returns plots that were not rejected.}
\item{TRUE}{returns all records}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1,3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots, such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as deer exclosures and bonus plots}}}
}
\value{
returns a dataframe with location and visit events
}
\description{
This function combines location and event data. Must run importData first.
}
\examples{
importCSV('./forest_csvs')
# Select most recent survey of data from APCO
APCO_data <- joinLocEvent(park = 'APCO', from = 2015, to = 2018)

# Select data from cycle 3
cycle3 <- joinLocEvent(from = 2015, to = 2018) # all parks is default

# Select data from plots that had a QA/QC event in GETT in 2018
GETT_data<-joinLocEvent(park = 'GETT', QAQC = T, from = 2018)
QAQC_plots<-GETT_data$Plot_Name[which(GETT_data$Event_QAQC==TRUE)]
GETT_QAQC<-GETT_data \%>\% filter(Plot_Name \%in\% QAQC_plots) \%>\% droplevels()

}
