% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRegenData.R
\name{joinRegenData}
\alias{joinRegenData}
\title{joinRegenData}
\usage{
joinRegenData(speciesType = c("all", "native", "exotic"),
  canopyForm = c("canopy", "all"), units = c("sq.m", "ha", "acres"),
  park = "all", from = 2007, to = 2018, QAQC = FALSE,
  locType = "VS", height = c("ht15", "all"), output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{canopyForm}{Allows you to filter on canopy species only or include all species.
\describe{
\item{"all"}{Returns all species, including low canopy species.}
\item{"canopy"}{Default. Returns canopy-forming species only}
}}

\item{units}{Calculates seedling and sapling densities based on different units.
\describe{
\item{"sq.m"}{Default. Returns seedling and sapling densities per square meter.}
\item{"ha"}{Returns seedling and sapling densities per hectare}
\item{"acres"}{Returns densities per acre}
}}

\item{height}{Determines whether all height classes are included, or just >15cm seedlings.
\describe{
\item{"ht15"}{Default. Returns seedlings >15cm and all saplings.}
\item{"all"}{Returns all seedlings and saplings, including 5-15cm seedlings in VAFO.}
}}
}
\value{
returns a dataframe with seedling and sapling densities, and stocking index
}
\description{
This function combines seedling and sapling data, and calculates stocking index. Must run importData first.
}
