% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumQuadGuilds.R
\name{sumQuadGuilds}
\alias{sumQuadGuilds}
\title{sumQuadGuilds: summarizes quadrat species data by guilds}
\usage{
sumQuadGuilds(
  speciesType = c("native", "exotic", "all"),
  park = "all",
  from = 2006,
  to = 2019,
  QAQC = FALSE,
  locType = "VS",
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Returns all species.}
\item{"native"}{Default. Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
Returns a dataframe with average quadrat cover, percent quadrat frequency and quadrat frequency count for tree,shrub/vine,herbaceous,and graminoid. Data are sither summarized for all species, native only, exotic only, or invasive only.
}
\description{
This function summarizes output from joinQuadData and calculates average cover and quadrat frequency for each guild. Average cover is corrected for number of quadrats sampled.
}
\examples{
importData()

# compile exotic quad data for all parks and most recent survey.
exo_guilds <- sumQuadGuilds(speciesType = 'exotic', from = 2015, to = 2018)

# compile native quad data for more recent survey in COLO, with ferns and forbs split in separate guilds
COLO_guilds <- sumQuadGuilds(speciesType = 'native', from = 2015, to = 2018)

}
