% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSapDBHDist.R
\name{sumSapDBHDist}
\alias{sumSapDBHDist}
\title{sumSapDBHDist: calculates DBH distribution of saplings}
\usage{
sumSapDBHDist(
  speciesType = c("all", "native", "exotic"),
  canopyForm = c("canopy", "all"),
  units = c("density", "ba"),
  park = "all",
  from = 2007,
  to = 2019,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{canopyForm}{Allows you to filter on canopy species only or include all species.
\describe{
\item{"all"}{Returns all species, including low canopy species.}
\item{"canopy"}{Default. Returns canopy-forming species only}
}}

\item{units}{Allows you to choose which metric to calculate: basal area or stem density
\describe{
\item{"density"}{Default. Returns stems/ha}
\item{"ba"}{Returns basal area in sq.m/ha}
}}
}
\value{
returns a dataframe with one row for each plot and either density or BA
}
\description{
This function calculates DBH distribution by 1cm size classes. Must run importData first.
}
\examples{
importData()
sap_diam_dist <- sumSapDBHDist(park = 'RICH', speciesType = 'native', from = 2015, to = 2018, units = 'ba')

}
