% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinQuadData.R
\name{joinQuadData}
\alias{joinQuadData}
\title{joinQuadData: compiles quadrat species data}
\usage{
joinQuadData(speciesType = c("all", "native", "exotic"), park = "all",
  from = 2007, to = 2018, QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
Returns a dataframe with cover class midpoints for each quadrat and includes guild for each species.
}
\description{
This function combines quadrat species and tree seedling cover data with species names and allows you to filter on species types, park, years, and visit type. Note that the Shrub guild also includes woody vine species.
}
\examples{
importData()
# compile quadrat data for exotic species in PETE for all years
PETE_quads <- joinQuadData(park = 'PETE', speciesType = 'exotic')

# compile native species only for all parks in most recent survey
native_quads <- joinQuadData(speciesType = 'native', from = 2015, to = 2018)

}
