% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRegenData.R
\name{joinRegenData}
\alias{joinRegenData}
\title{joinRegenData: compiles seedling and sapling data}
\usage{
joinRegenData(
  speciesType = c("all", "native", "exotic"),
  canopyForm = c("canopy", "all"),
  units = c("sq.m", "ha", "acres"),
  park = "all",
  from = 2007,
  to = 2019,
  QAQC = FALSE,
  locType = "VS",
  height = c("ht15", "all"),
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{canopyForm}{Allows you to filter on canopy species only or include all species.
\describe{
\item{"all"}{Returns all species, including low canopy species.}
\item{"canopy"}{Default. Returns canopy-forming species only}
}}

\item{units}{Calculates seedling and sapling densities based on different units.
\describe{
\item{"sq.m"}{Default. Returns seedling and sapling densities per square meter.}
\item{"ha"}{Returns seedling and sapling densities per hectare}
\item{"acres"}{Returns densities per acre}
}}

\item{height}{Determines whether all height classes are included, or just >15cm seedlings.
\describe{
\item{"ht15"}{Default. Returns seedlings >=15cm and all saplings.}
\item{"all"}{Returns all seedlings and saplings, including 5-15cm seedlings in VAFO.}
}}
}
\value{
returns a dataframe with seedling and sapling densities, stocking index,
quadrat seedling cover and quadrat seedling frequency. Quadrat frequency is based on
cover > 0 in a quadrat.
}
\description{
This function combines seedling and sapling data, and calculates stocking index at the 1 sq. m level.
Quadrat percent cover is also averaged across the quadrats, and only includes cover of seedling sized tree species
(i.e. the original protocol method). Stocking index thresholds are 2 for areas with low deer, and 8 for areas
with high deer. Must run importData first.
}
\examples{
importCSV('./forest_csvs/')
# Compile seedling and sapling data for all parks and all species in most recent cycle,
# and only include seedlings >=15cm tall (default).
regen_data <- joinRegenData(canopyForm = 'all', from = 2015, to = 2018)

# compile regen data for canopy-forming (default), native species of all size classes in VAFO for all years
VAFO_regen <- joinRegenData(park = 'VAFO', speciesType = 'native', height = 'all')

# Compile seedling and sapling densities as stems/ha for all parks in most recent survey
regen_data <- joinRegenData(units = 'ha', from = 2015, to = 2018)

}
