% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeData.R
\name{joinTreeData}
\alias{joinTreeData}
\title{joinTreeData: compiles tree data}
\usage{
joinTreeData(
  status = c("all", "live", "dead"),
  speciesType = c("all", "native", "exotic"),
  park = "all",
  from = 2007,
  to = 2019,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  dist_m = NA,
  output,
  ...
)
}
\arguments{
\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Includes all standing trees}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}
}
\value{
returns a dataframe with plot-level and visit-level tree data
}
\description{
This function combines location and event-level Tree data. Note that BA_cm2 was corrected
for Loc_Type=='Deer' to be the same as Loc_Type=='VS' by multiplying BA_cm2 by 4 to get BA_cm2/400m2.
Must run importData first.
}
\examples{
importData()
# compile tree data for live trees only in most recent survey in all parks
live_trees <- joinTreeData(status = 'live', from = 2015, to = 2018)

# compile FRSP trees within 7.3152m radius (FIA subplot size) in most recent survey
FRSP_100m <- joinTreeData(park = 'FRSP', from = 2015, to = 2018, dist_m = 7.3152)

# compile dead trees in GETT in most recent survey
GETT_dead <- joinTreeData(park = 'GETT', from = 2015, to = 2018, status = 'dead')

# compile exotic trees in VAFO in all years
VAFO_exotic <- joinTreeData(park = 'VAFO', from = 2015, to = 2018, speciesType = 'exotic')

}
